Refactor architecture: extract ViewModels, reusable views, native TabView
- Extract 4 reusable UI components (MtgLoadingView, MtgBadge, ColorPickerRow, ShareSheet)
  replacing duplicated patterns across 8+ view files
- Create 6 ViewModels (Profile, FriendsList, GameHistory, CreateGame, JoinGame, Home)
  moving business logic out of views into proper MVVM with protocol-based DI
- Split GameBoardView (998→387 lines) and LobbyView (564→255 lines) by extracting
  child views into separate files
- Split PlanechaseViews.swift (543 lines, 6 types) into Home/Planechase/ folder
  with one file per view
- Replace custom HStack tab bar with native TabView (4 tabs: Play, History,
  Friends, Profile), each with its own NavigationStack; tab bar hides during
  game sessions
