+++ Treachery-iOS/Treachery-iOS/Home/GameBoardView.swift
@@ -73,13 +73,7 @@ struct GameBoardView: View {
                 } else if viewModel.players.isEmpty {
                     // Loading state
                     Spacer()
-                    VStack(spacing: 12) {
-                        ProgressView()
-                            .tint(Color.mtgGold)
-                        Text("Loading game...")
-                            .font(.subheadline)
-                            .foregroundStyle(Color.mtgTextSecondary)
-                    }
+                    MtgLoadingView(message: "Loading game...")
                     Spacer()
                 } else {
                     // Identity card header — only when treachery active
@@ -163,7 +157,7 @@ struct GameBoardView: View {
 
                             VStack(spacing: 0) {
                                 ForEach(viewModel.players) { player in
-                                    PlayerRow(
+                                    GameBoardPlayerRow(
                                         player: player,
                                         viewModel: viewModel,
                                         showColorPicker: player.userId == viewModel.currentUserId ? $showColorPicker : .constant(false)
@@ -373,559 +367,6 @@ struct GameBoardView: View {
     }
 }
 
-// MARK: - Identity Card Header
-
-private struct IdentityCardHeader: View {
-    let card: IdentityCard
-    let player: Player
-
-    /// Leaders are always face-up; unveiled players show openly; everyone else is hidden.
-    private var isAlwaysVisible: Bool {
-        player.isUnveiled || player.role == .leader
-    }
-
-    var body: some View {
-        if isAlwaysVisible {
-            revealedContent
-        } else {
-            concealedContent
-        }
-    }
-
-    // MARK: - Concealed (tap to open sheet)
-
-    private var concealedContent: some View {
-        VStack(spacing: 8) {
-            HStack {
-                Image(systemName: "eye.slash.fill")
-                    .font(.title3)
-                    .foregroundStyle(Color.mtgGold)
-
-                Spacer()
-
-                HStack(spacing: 4) {
-                    Image(systemName: "heart.fill")
-                        .font(.caption)
-                        .foregroundStyle(Color.mtgError)
-                    Text("\(player.lifeTotal)")
-                        .font(.system(size: 28, weight: .bold, design: .serif))
-                        .foregroundStyle(Color.mtgTextPrimary)
-                        .contentTransition(.numericText())
-                }
-            }
-
-            Text("Tap to peek at your identity")
-                .font(.subheadline)
-                .foregroundStyle(Color.mtgTextSecondary)
-                .frame(maxWidth: .infinity, alignment: .leading)
-        }
-        .padding()
-        .background(Color.mtgSurface)
-        .overlay(
-            RoundedRectangle(cornerRadius: 0)
-                .stroke(Color.mtgBorderAccent, lineWidth: 1)
-        )
-    }
-
-    // MARK: - Revealed (unveiled)
-
-    private var revealedContent: some View {
-        VStack(spacing: 8) {
-            HStack {
-                HStack(spacing: 6) {
-                    Circle()
-                        .fill(player.role?.color ?? .gray)
-                        .frame(width: 12, height: 12)
-                        .shadow(color: (player.role?.color ?? .gray).opacity(0.5), radius: 4)
-                    Text(player.role?.displayName ?? "Unknown")
-                        .font(.headline)
-                        .foregroundStyle(player.role?.color ?? Color.mtgTextPrimary)
-                }
-
-                Spacer()
-
-                HStack(spacing: 4) {
-                    Image(systemName: "heart.fill")
-                        .font(.caption)
-                        .foregroundStyle(Color.mtgError)
-                    Text("\(player.lifeTotal)")
-                        .font(.system(size: 30, weight: .bold, design: .serif))
-                        .foregroundStyle(Color.mtgTextPrimary)
-                        .contentTransition(.numericText())
-                        .animation(.spring(response: 0.3, dampingFraction: 0.6), value: player.lifeTotal)
-                }
-            }
-
-            HStack {
-                Text(card.name)
-                    .font(.system(.title3, design: .serif))
-                    .fontWeight(.semibold)
-                    .foregroundStyle(Color.mtgTextPrimary)
-                Spacer()
-                Image(systemName: "chevron.right")
-                    .font(.caption)
-                    .foregroundStyle(Color.mtgTextSecondary)
-            }
-
-            Text(card.abilityText)
-                .font(.caption)
-                .foregroundStyle(Color.mtgTextSecondary)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .lineLimit(3)
-
-            HStack {
-                if player.isUnveiled {
-                    Text("UNVEILED")
-                        .font(.caption2)
-                        .fontWeight(.bold)
-                        .foregroundStyle(Color.mtgBackground)
-                        .padding(.horizontal, 8)
-                        .padding(.vertical, 2)
-                        .background(player.role?.color ?? .gray)
-                        .clipShape(Capsule())
-                } else if player.role == .leader {
-                    Text("LEADER — ALWAYS VISIBLE")
-                        .font(.caption2)
-                        .fontWeight(.bold)
-                        .foregroundStyle(Color.mtgGold)
-                        .padding(.horizontal, 8)
-                        .padding(.vertical, 2)
-                        .background(Color.mtgGold.opacity(0.15))
-                        .clipShape(Capsule())
-                }
-                Spacer()
-                Text("Tap for details")
-                    .font(.caption2)
-                    .foregroundStyle(Color.mtgTextSecondary)
-            }
-        }
-        .padding()
-        .background(
-            ZStack {
-                Color.mtgSurface
-                LinearGradient(
-                    colors: [
-                        (player.role?.color ?? Color.mtgGold).opacity(0.1),
-                        Color.clear
-                    ],
-                    startPoint: .topLeading,
-                    endPoint: .bottomTrailing
-                )
-            }
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 0)
-                .stroke(player.role?.color ?? Color.mtgBorderAccent, lineWidth: 1)
-        )
-    }
-}
-
-// MARK: - Player Row
-
-private struct PlayerRow: View {
-    let player: Player
-    @ObservedObject var viewModel: GameBoardViewModel
-    @Binding var showColorPicker: Bool
-    var onViewCard: ((Player) -> Void)?
-
-    private var isCurrentUser: Bool {
-        player.userId == viewModel.currentUserId
-    }
-
-    /// Whether this player's card can be inspected by the current user.
-    /// True for unveiled players and leaders (but not yourself -- you have the header).
-    private var canInspectCard: Bool {
-        guard !isCurrentUser else { return false }
-        guard viewModel.identityCard(for: player) != nil else { return false }
-        return player.isUnveiled || player.role == .leader
-    }
-
-    var body: some View {
-        VStack(spacing: 0) {
-            HStack(spacing: 0) {
-                // Left accent bar for player color
-                if let hex = player.playerColor {
-                    RoundedRectangle(cornerRadius: 1.5)
-                        .fill(Color(hex: hex))
-                        .frame(width: 3)
-                        .padding(.vertical, 2)
-                        .padding(.trailing, 8)
-                }
-
-                // Color picker toggle for current user
-                if isCurrentUser {
-                    Button {
-                        withAnimation(.easeInOut(duration: 0.2)) {
-                            showColorPicker.toggle()
-                        }
-                    } label: {
-                        if let hex = player.playerColor {
-                            Circle()
-                                .fill(Color(hex: hex))
-                                .frame(width: 16, height: 16)
-                                .overlay(
-                                    Circle().stroke(Color.mtgTextSecondary.opacity(0.3), lineWidth: 1)
-                                )
-                        } else {
-                            Circle()
-                                .stroke(Color.mtgTextSecondary, lineWidth: 1.5)
-                                .frame(width: 16, height: 16)
-                        }
-                    }
-                    .buttonStyle(.plain)
-                    .padding(.trailing, 6)
-                    .accessibilityLabel("Choose player color")
-                }
-
-                // Player info
-                VStack(alignment: .leading, spacing: 2) {
-                    HStack(spacing: 6) {
-                        Text(player.displayName)
-                            .fontWeight(isCurrentUser ? .bold : .regular)
-                            .strikethrough(player.isEliminated)
-                            .foregroundStyle(player.isEliminated ? Color.mtgTextSecondary : Color.mtgTextPrimary)
-
-                        if isCurrentUser {
-                            Text("You")
-                                .font(.caption2)
-                                .foregroundStyle(Color.mtgGold)
-                                .padding(.horizontal, 6)
-                                .padding(.vertical, 1)
-                                .background(Color.mtgGold.opacity(0.15))
-                                .clipShape(Capsule())
-                        }
-
-                        if player.isEliminated {
-                            Image(systemName: "xmark.circle.fill")
-                                .foregroundStyle(Color.mtgError)
-                                .font(.caption)
-                        }
-                    }
-
-                    // Commander name
-                    if let commanderName = player.commanderName, !commanderName.isEmpty {
-                        Text(commanderName)
-                            .font(.system(.caption, design: .serif))
-                            .italic()
-                            .foregroundStyle(Color.mtgTextSecondary)
-                    }
-
-                    // Role visibility — only show for unveiled players and leaders
-                    // to prevent leaking info when the phone is on the table
-                    if player.isUnveiled || player.role == .leader {
-                        if canInspectCard {
-                            Button {
-                                onViewCard?(player)
-                            } label: {
-                                HStack(spacing: 4) {
-                                    Circle()
-                                        .fill(player.role?.color ?? .gray)
-                                        .frame(width: 8, height: 8)
-                                    Text(player.role?.displayName ?? "")
-                                        .font(.caption)
-                                        .foregroundStyle(player.role?.color ?? Color.mtgTextSecondary)
-                                    if player.isUnveiled && player.role != .leader {
-                                        Text("(Unveiled)")
-                                            .font(.caption2)
-                                            .foregroundStyle(Color.mtgTextSecondary)
-                                    }
-                                    Image(systemName: "info.circle")
-                                        .font(.caption2)
-                                        .foregroundStyle(player.role?.color ?? Color.mtgTextSecondary)
-                                }
-                            }
-                            .buttonStyle(.plain)
-                            .accessibilityLabel("View \(player.displayName)'s identity card")
-                            .accessibilityHint("Shows their role ability and card details")
-                        } else {
-                            HStack(spacing: 4) {
-                                Circle()
-                                    .fill(player.role?.color ?? .gray)
-                                    .frame(width: 8, height: 8)
-                                Text(player.role?.displayName ?? "")
-                                    .font(.caption)
-                                    .foregroundStyle(player.role?.color ?? Color.mtgTextSecondary)
-                                if player.isUnveiled && player.role != .leader {
-                                    Text("(Unveiled)")
-                                        .font(.caption2)
-                                        .foregroundStyle(Color.mtgTextSecondary)
-                                }
-                            }
-                        }
-                    } else if viewModel.isTreacheryActive {
-                        Text("Role Hidden")
-                            .font(.caption)
-                            .foregroundStyle(Color.mtgTextSecondary)
-                    }
-                }
-
-                Spacer()
-
-                // Life controls
-                if !player.isEliminated {
-                    HStack(spacing: 12) {
-                        Button {
-                            viewModel.adjustLife(for: player.id, by: -1)
-                        } label: {
-                            Image(systemName: "minus.circle.fill")
-                                .font(.system(size: 32))
-                                .foregroundStyle(Color.mtgAssassin)
-                                .background(
-                                    Circle()
-                                        .fill(Color.mtgAssassin.opacity(0.1))
-                                        .frame(width: 40, height: 40)
-                                )
-                        }
-                        .buttonStyle(.plain)
-                        .accessibilityLabel("Decrease \(player.displayName)'s life")
-
-                        Text("\(player.lifeTotal)")
-                            .font(.system(size: 36, weight: .bold, design: .serif))
-                            .foregroundStyle(Color.mtgTextPrimary)
-                            .frame(minWidth: 52)
-                            .multilineTextAlignment(.center)
-                            .contentTransition(.numericText())
-                            .animation(.spring(response: 0.3, dampingFraction: 0.6), value: player.lifeTotal)
-                            .accessibilityLabel("\(player.lifeTotal) life")
-
-                        Button {
-                            viewModel.adjustLife(for: player.id, by: 1)
-                        } label: {
-                            Image(systemName: "plus.circle.fill")
-                                .font(.system(size: 32))
-                                .foregroundStyle(Color.mtgSuccess)
-                                .background(
-                                    Circle()
-                                        .fill(Color.mtgSuccess.opacity(0.1))
-                                        .frame(width: 40, height: 40)
-                                )
-                        }
-                        .buttonStyle(.plain)
-                        .accessibilityLabel("Increase \(player.displayName)'s life")
-                    }
-                } else {
-                    Text("Eliminated")
-                        .font(.caption)
-                        .foregroundStyle(Color.mtgError)
-                }
-            }
-            .padding(.vertical, 4)
-
-            // Color picker row (current user only)
-            if isCurrentUser && showColorPicker {
-                ColorPickerRow(selectedHex: player.playerColor) { hex in
-                    Task { await viewModel.updatePlayerColor(hex) }
-                }
-                .padding(.top, 8)
-                .transition(.asymmetric(
-                    insertion: .move(edge: .top).combined(with: .opacity),
-                    removal: .opacity
-                ))
-            }
-        }
-        .accessibilityElement(children: .combine)
-    }
-}
-
-// MARK: - Color Picker Row
-
-private struct ColorPickerRow: View {
-    let selectedHex: String?
-    let onSelect: (String?) -> Void
-
-    var body: some View {
-        HStack(spacing: 8) {
-            ForEach(PlayerColors.palette, id: \.hex) { playerColor in
-                Button {
-                    if selectedHex == playerColor.hex {
-                        onSelect(nil)
-                    } else {
-                        onSelect(playerColor.hex)
-                    }
-                } label: {
-                    Circle()
-                        .fill(playerColor.color)
-                        .frame(width: 24, height: 24)
-                        .overlay(
-                            Circle()
-                                .stroke(Color.mtgTextPrimary, lineWidth: selectedHex == playerColor.hex ? 2 : 0)
-                                .padding(selectedHex == playerColor.hex ? -2 : 0)
-                        )
-                }
-                .buttonStyle(.plain)
-                .accessibilityLabel(playerColor.name)
-            }
-
-            // Clear button
-            Button {
-                onSelect(nil)
-            } label: {
-                ZStack {
-                    Circle()
-                        .stroke(Color.mtgTextSecondary, lineWidth: 1)
-                        .frame(width: 24, height: 24)
-                    Image(systemName: "xmark")
-                        .font(.system(size: 10, weight: .bold))
-                        .foregroundStyle(Color.mtgTextSecondary)
-                }
-            }
-            .buttonStyle(.plain)
-            .accessibilityLabel("Clear color")
-        }
-        .padding(.vertical, 4)
-    }
-}
-
-// MARK: - Action Bar
-
-private struct ActionBar: View {
-    @ObservedObject var viewModel: GameBoardViewModel
-    @State private var showUnveilConfirmation = false
-    var onUnveil: () -> Void
-
-    var body: some View {
-        VStack(spacing: 12) {
-            if let player = viewModel.currentPlayer,
-               !player.isUnveiled,
-               !player.isEliminated,
-               player.role != .leader {
-                Button("Unveil Identity") {
-                    showUnveilConfirmation = true
-                }
-                .buttonStyle(MtgPrimaryButtonStyle(isDisabled: viewModel.isPending))
-                .disabled(viewModel.isPending)
-                .padding(.horizontal)
-                .accessibilityLabel("Unveil your identity as \(player.role?.displayName ?? "unknown")")
-                .accessibilityHint("Reveals your role to all players. Cannot be undone.")
-                .confirmationDialog(
-                    "Unveil your identity?",
-                    isPresented: $showUnveilConfirmation,
-                    titleVisibility: .visible
-                ) {
-                    Button("Unveil") {
-                        onUnveil()
-                    }
-                    Button("Cancel", role: .cancel) {}
-                } message: {
-                    if let card = viewModel.currentIdentityCard {
-                        Text("This will reveal your role (\(player.role?.displayName ?? "")) and card (\(card.name)) to all players. This cannot be undone.")
-                    }
-                }
-            }
-
-        }
-        .padding()
-    }
-}
-
-// MARK: - Winner Selection Sheet
-
-private struct WinnerSelectionSheet: View {
-    let players: [Player]
-    @Binding var selectedWinners: Set<String>
-    let isPending: Bool
-    let onConfirm: () -> Void
-    let onCancel: () -> Void
-
-    var body: some View {
-        NavigationStack {
-            ZStack {
-                Color.mtgBackground.ignoresSafeArea()
-
-                VStack(spacing: 16) {
-                    Text("Select the winner(s) of this game for ELO tracking.")
-                        .font(.subheadline)
-                        .foregroundStyle(Color.mtgTextSecondary)
-                        .multilineTextAlignment(.center)
-                        .padding(.horizontal)
-
-                    ScrollView {
-                        VStack(spacing: 0) {
-                            ForEach(players) { player in
-                                Button {
-                                    if selectedWinners.contains(player.userId) {
-                                        selectedWinners.remove(player.userId)
-                                    } else {
-                                        selectedWinners.insert(player.userId)
-                                    }
-                                } label: {
-                                    HStack(spacing: 12) {
-                                        Image(systemName: selectedWinners.contains(player.userId) ? "checkmark.circle.fill" : "circle")
-                                            .foregroundStyle(selectedWinners.contains(player.userId) ? Color.mtgSuccess : Color.mtgTextSecondary)
-                                            .font(.title3)
-
-                                        VStack(alignment: .leading, spacing: 2) {
-                                            Text(player.displayName)
-                                                .foregroundStyle(Color.mtgTextPrimary)
-                                            if let commanderName = player.commanderName, !commanderName.isEmpty {
-                                                Text(commanderName)
-                                                    .font(.system(.caption, design: .serif))
-                                                    .italic()
-                                                    .foregroundStyle(Color.mtgTextSecondary)
-                                            }
-                                        }
-
-                                        Spacer()
-
-                                        if let hex = player.playerColor {
-                                            Circle()
-                                                .fill(Color(hex: hex))
-                                                .frame(width: 12, height: 12)
-                                        }
-                                    }
-                                    .padding(.horizontal, 16)
-                                    .padding(.vertical, 12)
-                                }
-                                .buttonStyle(.plain)
-
-                                if player.id != players.last?.id {
-                                    Rectangle()
-                                        .fill(Color.mtgDivider)
-                                        .frame(height: 1)
-                                        .padding(.horizontal, 16)
-                                }
-                            }
-                        }
-                        .mtgCardFrame()
-                        .padding(.horizontal)
-                    }
-
-                    VStack(spacing: 8) {
-                        Button {
-                            onConfirm()
-                        } label: {
-                            if isPending {
-                                HStack(spacing: 8) {
-                                    ProgressView().controlSize(.small).tint(Color.mtgBackground)
-                                    Text("Ending Game...")
-                                }
-                            } else {
-                                Text("End Game")
-                            }
-                        }
-                        .buttonStyle(MtgPrimaryButtonStyle(isDisabled: isPending))
-                        .disabled(isPending)
-                        .padding(.horizontal)
-
-                        Text("You can skip winner selection — ELO won't be updated.")
-                            .font(.caption2)
-                            .foregroundStyle(Color.mtgTextSecondary)
-                    }
-                    .padding(.bottom)
-                }
-            }
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarColorScheme(.dark, for: .navigationBar)
-            .toolbar {
-                ToolbarItem(placement: .topBarLeading) {
-                    Button("Cancel") { onCancel() }
-                        .foregroundStyle(Color.mtgGold)
-                }
-            }
-        }
-    }
-}
-
 // MARK: - Previews
 
 #if DEBUG
@@ -942,57 +383,5 @@ private struct WinnerSelectionSheet: View {
         )
     }
     .environmentObject(AuthViewModel())
-}
-
-#Preview("Identity Card Header - Leader (always visible)") {
-    IdentityCardHeader(
-        card: .sampleLeaderCard,
-        player: .sampleLeader
-    )
-}
-
-#Preview("Identity Card Header - Hidden (tap to peek)") {
-    IdentityCardHeader(
-        card: .sampleGuardianCard,
-        player: .sampleGuardian
-    )
-}
-
-#Preview("Identity Card Header - Unveiled") {
-    IdentityCardHeader(
-        card: .sampleAssassinCard,
-        player: .sampleAssassin
-    )
-}
-
-#Preview("Player Row - Active") {
-    let vm = GameBoardViewModel(
-        gameId: "preview",
-        previewPlayers: Player.sampleGamePlayers,
-        previewGame: .sampleInProgress,
-        currentUserId: "user2"
-    )
-    ZStack {
-        Color.mtgBackground.ignoresSafeArea()
-        VStack(spacing: 0) {
-            PlayerRow(
-                player: .sampleGuardian,
-                viewModel: vm,
-                showColorPicker: .constant(false)
-            )
-            PlayerRow(
-                player: .sampleAssassin,
-                viewModel: vm,
-                showColorPicker: .constant(false)
-            )
-            PlayerRow(
-                player: .sampleEliminated,
-                viewModel: vm,
-                showColorPicker: .constant(false)
-            )
-        }
-        .mtgCardFrame()
-        .padding()
-    }
 }
 #endif
