+++ Treachery-iOS/Treachery-iOS/Home/LobbyView.swift
@@ -93,9 +93,7 @@ struct LobbyView: View {
     private var loadingContent: some View {
         VStack {
             Spacer()
-            ProgressView("Loading lobby...")
-                .tint(Color.mtgGold)
-                .foregroundStyle(Color.mtgTextSecondary)
+            MtgLoadingView(message: "Loading lobby...")
             Spacer()
         }
     }
@@ -238,313 +236,6 @@ struct LobbyView: View {
             }
         }
     }
-}
-
-// MARK: - Game Code Card
-
-private struct LobbyGameCodeCard: View {
-    let game: Game
-    @Binding var showShareSheet: Bool
-
-    var body: some View {
-        VStack(spacing: 10) {
-            MtgSectionHeader(title: "Game Code")
-
-            Text(game.code)
-                .font(.system(size: 52, weight: .bold, design: .monospaced))
-                .mtgGoldShimmer()
-                .kerning(10)
-                .shadow(color: Color.mtgGold.opacity(0.3), radius: 12, x: 0, y: 0)
-                .accessibilityLabel("Game code: \(game.code.map(String.init).joined(separator: " "))")
-
-            Text(game.gameMode.displayName)
-                .font(.caption)
-                .fontWeight(.semibold)
-                .foregroundStyle(Color.mtgBackground)
-                .padding(.horizontal, 12)
-                .padding(.vertical, 4)
-                .background(
-                    LinearGradient(
-                        colors: [Color(hex: "e4c96a"), Color(hex: "c9a84c")],
-                        startPoint: .leading,
-                        endPoint: .trailing
-                    )
-                )
-                .clipShape(Capsule())
-                .accessibilityLabel("Game mode: \(game.gameMode.displayName)")
-
-            HStack(spacing: 16) {
-                Button {
-                    UIPasteboard.general.string = game.code
-                } label: {
-                    HStack(spacing: 4) {
-                        Image(systemName: "doc.on.doc")
-                            .font(.caption)
-                        Text("Copy")
-                            .font(.caption)
-                    }
-                    .foregroundStyle(Color.mtgGold)
-                }
-                .accessibilityLabel("Copy game code")
-
-                Button {
-                    showShareSheet = true
-                } label: {
-                    HStack(spacing: 4) {
-                        Image(systemName: "square.and.arrow.up")
-                            .font(.caption)
-                        Text("Share")
-                            .font(.caption)
-                    }
-                    .foregroundStyle(Color.mtgGold)
-                }
-                .accessibilityLabel("Share game code with friends")
-            }
-            .padding(.top, 4)
-        }
-        .padding(.vertical, 24)
-        .padding(.horizontal, 20)
-        .mtgCardFrame()
-        .mtgCardGlow(color: .mtgGold, radius: 12, opacity: 0.12)
-        .padding(.horizontal)
-        .padding(.top, 12)
-    }
-}
-
-// MARK: - Player Row
-
-private struct LobbyPlayerRow: View {
-    let player: Player
-    let isMe: Bool
-    let isHost: Bool
-    @Binding var showColorPicker: Bool
-    @Binding var commanderNameInput: String
-    var onColorChange: (String?) -> Void
-    var onCommanderNameChange: (String) -> Void
-    var onCommanderNameSubmit: (String) -> Void
-
-    var body: some View {
-        VStack(spacing: 0) {
-            HStack(spacing: 0) {
-                if let hex = player.playerColor {
-                    RoundedRectangle(cornerRadius: 1.5)
-                        .fill(Color(hex: hex))
-                        .frame(width: 3)
-                        .padding(.vertical, 2)
-                        .padding(.trailing, 8)
-                }
-
-                if isMe {
-                    colorPickerToggle
-                }
-
-                VStack(alignment: .leading, spacing: 2) {
-                    Text(player.displayName)
-                        .fontWeight(isHost ? .semibold : .regular)
-                        .foregroundStyle(Color.mtgTextPrimary)
-
-                    if !isMe, let commanderName = player.commanderName, !commanderName.isEmpty {
-                        Text(commanderName)
-                            .font(.system(.caption, design: .serif))
-                            .italic()
-                            .foregroundStyle(Color.mtgTextSecondary)
-                    }
-                }
-
-                Spacer()
-
-                if isHost {
-                    Text("Host")
-                        .font(.caption)
-                        .foregroundStyle(Color.mtgGold)
-                        .padding(.horizontal, 8)
-                        .padding(.vertical, 2)
-                        .background(Color.mtgGold.opacity(0.15))
-                        .clipShape(Capsule())
-                }
-            }
-
-            if isMe {
-                commanderNameField
-            }
-
-            if isMe && showColorPicker {
-                LobbyColorPickerRow(selectedHex: player.playerColor) { hex in
-                    onColorChange(hex)
-                }
-                .padding(.top, 8)
-                .transition(.asymmetric(
-                    insertion: .move(edge: .top).combined(with: .opacity),
-                    removal: .opacity
-                ))
-            }
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 14)
-        .accessibilityLabel("\(player.displayName)\(isHost ? ", Host" : "")")
-    }
-
-    private var colorPickerToggle: some View {
-        Button {
-            withAnimation(.easeInOut(duration: 0.2)) {
-                showColorPicker.toggle()
-            }
-        } label: {
-            if let hex = player.playerColor {
-                Circle()
-                    .fill(Color(hex: hex))
-                    .frame(width: 16, height: 16)
-                    .overlay(
-                        Circle().stroke(Color.mtgTextSecondary.opacity(0.3), lineWidth: 1)
-                    )
-            } else {
-                Circle()
-                    .stroke(Color.mtgTextSecondary, lineWidth: 1.5)
-                    .frame(width: 16, height: 16)
-            }
-        }
-        .buttonStyle(.plain)
-        .padding(.trailing, 6)
-        .accessibilityLabel("Choose player color")
-    }
-
-    private var commanderNameField: some View {
-        TextField("Commander name...", text: $commanderNameInput)
-            .font(.system(.caption, design: .serif))
-            .italic()
-            .foregroundStyle(Color.mtgTextPrimary)
-            .padding(.horizontal, 10)
-            .padding(.vertical, 6)
-            .background(Color.mtgCardElevated)
-            .clipShape(RoundedRectangle(cornerRadius: 6))
-            .overlay(
-                RoundedRectangle(cornerRadius: 6)
-                    .stroke(Color.mtgDivider, lineWidth: 1)
-            )
-            .padding(.top, 6)
-            .onSubmit {
-                onCommanderNameSubmit(commanderNameInput)
-            }
-            .onChange(of: commanderNameInput) { _, newValue in
-                onCommanderNameChange(newValue)
-            }
-    }
-}
-
-// MARK: - Bottom Buttons
-
-private struct LobbyBottomButtons: View {
-    @ObservedObject var viewModel: LobbyViewModel
-    @Binding var isLeaving: Bool
-    var onLeave: () -> Void
-
-    var body: some View {
-        VStack(spacing: 12) {
-            if viewModel.isHost {
-                Button {
-                    Task { await viewModel.startGame() }
-                } label: {
-                    if viewModel.isStartingGame {
-                        HStack(spacing: 8) {
-                            ProgressView()
-                                .controlSize(.small)
-                                .tint(Color.mtgBackground)
-                            Text("Starting...")
-                        }
-                    } else {
-                        Text("Start Game")
-                    }
-                }
-                .buttonStyle(MtgPrimaryButtonStyle(isDisabled: !viewModel.canStartGame || viewModel.isStartingGame))
-                .disabled(!viewModel.canStartGame || viewModel.isStartingGame)
-
-                if !viewModel.canStartGame && viewModel.players.count < viewModel.minimumPlayerCount {
-                    Text("Need at least \(viewModel.minimumPlayerCount) player\(viewModel.minimumPlayerCount == 1 ? "" : "s") to start")
-                        .font(.caption)
-                        .foregroundStyle(Color.mtgTextSecondary)
-                }
-            }
-
-            Button {
-                onLeave()
-            } label: {
-                if isLeaving {
-                    HStack(spacing: 8) {
-                        ProgressView()
-                            .controlSize(.small)
-                            .tint(Color.mtgError)
-                        Text("Leaving...")
-                    }
-                } else {
-                    Text("Leave Game")
-                }
-            }
-            .foregroundStyle(Color.mtgError)
-            .disabled(isLeaving)
-            .accessibilityLabel("Leave game")
-        }
-        .padding()
-    }
-}
-
-// MARK: - Lobby Color Picker Row
-
-private struct LobbyColorPickerRow: View {
-    let selectedHex: String?
-    let onSelect: (String?) -> Void
-
-    var body: some View {
-        HStack(spacing: 8) {
-            ForEach(PlayerColors.palette, id: \.hex) { playerColor in
-                Button {
-                    if selectedHex == playerColor.hex {
-                        onSelect(nil)
-                    } else {
-                        onSelect(playerColor.hex)
-                    }
-                } label: {
-                    Circle()
-                        .fill(playerColor.color)
-                        .frame(width: 24, height: 24)
-                        .overlay(
-                            Circle()
-                                .stroke(Color.mtgTextPrimary, lineWidth: selectedHex == playerColor.hex ? 2 : 0)
-                                .padding(selectedHex == playerColor.hex ? -2 : 0)
-                        )
-                }
-                .buttonStyle(.plain)
-                .accessibilityLabel(playerColor.name)
-            }
-
-            Button {
-                onSelect(nil)
-            } label: {
-                ZStack {
-                    Circle()
-                        .stroke(Color.mtgTextSecondary, lineWidth: 1)
-                        .frame(width: 24, height: 24)
-                    Image(systemName: "xmark")
-                        .font(.system(size: 10, weight: .bold))
-                        .foregroundStyle(Color.mtgTextSecondary)
-                }
-            }
-            .buttonStyle(.plain)
-            .accessibilityLabel("Clear color")
-        }
-        .padding(.vertical, 4)
-    }
-}
-
-// MARK: - Share Sheet
-
-private struct ShareSheet: UIViewControllerRepresentable {
-    let items: [Any]
-
-    func makeUIViewController(context: Context) -> UIActivityViewController {
-        UIActivityViewController(activityItems: items, applicationActivities: nil)
-    }
-
-    func updateUIViewController(_ uiViewController: UIActivityViewController, context: Context) {}
 }
 
 #if DEBUG
