+++ Treachery-iOS/Treachery-iOS/Views/ShareSheet.swift
@@ -1,0 +1,16 @@
+//
+//  ShareSheet.swift
+//  Treachery-iOS
+//
+
+import SwiftUI
+
+struct ShareSheet: UIViewControllerRepresentable {
+    let items: [Any]
+
+    func makeUIViewController(context: Context) -> UIActivityViewController {
+        UIActivityViewController(activityItems: items, applicationActivities: nil)
+    }
+
+    func updateUIViewController(_ uiViewController: UIActivityViewController, context: Context) {}
+}
